refactor: extract ProfileDataStore from UserDataStore
