+++ app-shared/src/iosMain/kotlin/io/github/droidkaigi/confsched/IosAppGraph.kt
@@ -17,6 +17,7 @@ import dev.zacsweers.metro.createGraph
 import io.github.droidkaigi.confsched.common.scope.TimetableDetailScope
 import io.github.droidkaigi.confsched.data.DataScope
 import io.github.droidkaigi.confsched.data.DataStoreDependencyProviders
+import io.github.droidkaigi.confsched.data.ProfileDataStoreQualifier
 import io.github.droidkaigi.confsched.data.SessionCacheDataStoreQualifier
 import io.github.droidkaigi.confsched.data.SettingsDataStoreQualifier
 import io.github.droidkaigi.confsched.data.UserDataStoreQualifier
@@ -230,6 +231,21 @@ interface IosAppGraph : AppGraph {
             migrations = emptyList(),
             scope = CoroutineScope(ioDispatcher + SupervisorJob()),
             produceFile = { dataStorePathProducer.producePath(DataStoreDependencyProviders.DATA_STORE_SETTINGS_FILE_NAME).toPath() },
+        )
+    }
+
+    @SingleIn(DataScope::class)
+    @ProfileDataStoreQualifier
+    @Provides
+    fun provideProfileDataStore(
+        dataStorePathProducer: DataStorePathProducer,
+        @IoDispatcher ioDispatcher: CoroutineDispatcher,
+    ): DataStore<Preferences> {
+        return PreferenceDataStoreFactory.createWithPath(
+            corruptionHandler = ReplaceFileCorruptionHandler({ emptyPreferences() }),
+            migrations = emptyList(),
+            scope = CoroutineScope(ioDispatcher + SupervisorJob()),
+            produceFile = { dataStorePathProducer.producePath(DataStoreDependencyProviders.DATA_STORE_PROFILE_FILE_NAME).toPath() },
         )
     }
