+++ core/data/src/commonMain/kotlin/io/github/droidkaigi/confsched/data/profile/DefaultProfileMutationKey.kt
@@ -11,7 +11,7 @@ import soil.query.buildMutationKey
 @ContributesBinding(DataScope::class)
 @Inject
 public class DefaultProfileMutationKey(
-    private val dataStore: UserDataStore,
+    private val dataStore: ProfileDataStore,
 ) : ProfileMutationKey by buildMutationKey(
     id = MutationId("profile_mutation_key"),
     mutate = { dataStore.saveProfile(it) },
