+++ core/data/src/commonMain/kotlin/io/github/droidkaigi/confsched/data/user/UserDataStore.kt
@@ -5,13 +5,11 @@ import androidx.datastore.core.IOException
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.emptyPreferences
-import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.datastore.preferences.core.stringSetPreferencesKey
 import dev.zacsweers.metro.Inject
 import dev.zacsweers.metro.SingleIn
 import io.github.droidkaigi.confsched.data.DataScope
 import io.github.droidkaigi.confsched.data.UserDataStoreQualifier
-import io.github.droidkaigi.confsched.model.profile.Profile
 import io.github.droidkaigi.confsched.model.sessions.TimetableItemId
 import kotlinx.collections.immutable.PersistentSet
 import kotlinx.collections.immutable.toPersistentSet
@@ -19,13 +17,11 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.map
-import kotlinx.serialization.json.Json
 
 @SingleIn(DataScope::class)
 @Inject
 public class UserDataStore(
     @param:UserDataStoreQualifier private val dataStore: DataStore<Preferences>,
-    private val json: Json,
 ) {
     public fun getStream(): Flow<PersistentSet<TimetableItemId>> {
         return dataStore.data
@@ -55,23 +51,6 @@ public class UserDataStore(
         }
     }
 
-    public fun getProfileOrNull(): Flow<Profile?> {
-        return dataStore.data.map {
-            val profile = it[PROFILE_KEY] ?: return@map null
-            try {
-                json.decodeFromString(profile)
-            } catch (_: Throwable) {
-                null
-            }
-        }
-    }
-
-    public suspend fun saveProfile(profile: Profile) {
-        dataStore.edit { preferences ->
-            preferences[PROFILE_KEY] = json.encodeToString(Profile.serializer(), profile)
-        }
-    }
-
     public suspend fun clearFavorites() {
         dataStore.edit { preferences ->
             preferences.remove(FAVORITE_SESSION_IDS_KEY)
@@ -80,6 +59,5 @@ public class UserDataStore(
 
     private companion object {
         private val FAVORITE_SESSION_IDS_KEY = stringSetPreferencesKey("FAVORITE_SESSION_IDS_KEY")
-        private val PROFILE_KEY = stringPreferencesKey("PROFILE_KEY")
     }
 }
